% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{cleaning_data}
\alias{cleaning_data}
\title{Clean the initial dataframe}
\usage{
cleaning_data(input_data)
}
\arguments{
\item{input_data}{A dataframe that contains columns 'Date', 'Tmin', 'Tmax', 'PPT'}
}
\value{
A dataframe with date columns of class Date, and the other variables as numeric
}
\description{
Clean the initial dataframe
}
\examples{
cleaning_data(input_data)
}
