\name{getParameters.fitContinuous}
\alias{getParameters.fitContinuous}
\title{Method to extract the parameters}
\usage{
  \method{getParameters}{fitContinuous} (fit)
}
\arguments{
  \item{fit}{a fitContinuous object}
}
\value{
  a list of parameters, currently only for a single/first
  trait
}
\description{
  Method to extract the parameters
}
\keyword{internal}

