\name{loglik.fitContinuous}
\alias{loglik.fitContinuous}
\title{Method to extract the log likelihood}
\usage{
  \method{loglik}{fitContinuous} (fit)
}
\arguments{
  \item{fit}{a fitContinuous object}
}
\value{
  the log likelihood, currently only for a single/first
  trait
}
\description{
  Method to extract the log likelihood
}
\keyword{internal}

