% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mildred_Lake}
\alias{Mildred_Lake}
\title{Mildred Lake Dataset}
\format{
A tibble (or data frame) with 730 rows and 5 variables:
\describe{
\item{Date}{Date of the observation (YYYY-MM-DD)}
\item{Tmin}{Minimum daily temperature (°C)}
\item{Tmax}{Maximum daily temperature (°C)}
\item{PPT}{Daily precipitation (cm)}
}
}
\source{
Data obtained from \href{https://climatedata.ca/download/#station-download}{ClimateData.ca}.
}
\usage{
data(Mildred_Lake)
}
\description{
A dataset that includes daily weather observations collected from the Mildred Lake weather station in Alberta.
\strong{Location Coordinates:} Latitude: 57.0833, Longitude: -111.5833.
}
\keyword{datasets}
