% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_to_cm.R
\name{mm_to_cm}
\alias{mm_to_cm}
\title{Function to convert a variable from millimeters (mm) to centimeters (cm)}
\usage{
mm_to_cm(input_data, columns)
}
\arguments{
\item{columns}{Selected column by the user such as 'PET' and/or 'PPT'.}

\item{input.data}{A dataframe that contains columns 'Date', 'Tmin', 'Tmax', 'PPT'}
}
\value{
A dataframe with original columns, except for the selected columns converted to centimeters.
}
\description{
Function to convert a variable from millimeters (mm) to centimeters (cm)
}
\examples{
mm_to_cm(input_data, columns = c("PPT"))
}
