\name{plot.pmc}
\alias{plot.pmc}
\title{plot the distributions}
\usage{
  \method{plot}{pmc} (x, ...)
}
\arguments{
  \item{x}{a pmc object}

  \item{...}{Additional arguments: A= a name for the first
  model in the pmc pairwise comparison B= a name for the
  second model in the pairwise comparison}
}
\description{
  plot the distributions
}

