% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc-package.R
\docType{package}
\name{pmc-package}
\alias{pmc-package}
\alias{_PACKAGE}
\title{pmc: Phylogenetic Monte Carlo}
\description{
Monte Carlo based model choice for applied phylogenetics of continuous traits. Method described in Carl Boettiger, Graham Coop, Peter Ralph (2012) Is your phylogeny informative? Measuring the power of comparative methods, Evolution 66 (7) 2240-51. \doi{10.1111/j.1558-5646.2011.01574.x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cboettig/pmc}
  \item Report bugs at \url{https://github.com/cboettig/pmc/issues}
}

}
\author{
\strong{Maintainer}: Carl Boettiger \email{cboettig@gmail.com}

}
\keyword{internal}
