% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMCISI_calc.R
\name{PMCISI_calc}
\alias{PMCISI_calc}
\title{Peat Spread Rate Index}
\usage{
PMCISI_calc(input_data, year_to_plot = NULL)
}
\arguments{
\item{input_data}{A dataframe containing the columns 'Date', 'PMC' and 'ISI'.}

\item{year_to_plot}{The year to plot.}
}
\value{
A dataframe that retains the original columns and adds a column with the calculated PMC_ISI.
}
\description{
The PMCISI combines the ISI and PMC to provide a representation of
above-ground fire spread potential in peatland fuels.
}
\examples{
input_data <- data.frame(Date = as.Date(c("2014-01-01", "2014-01-02",
"2014-01-03")) , PMC = c(10.005600, 9.686799, 9.688504),
ISI = c(0.2, 0.1, 0.1))
PMCISI_calc(input_data, year_to_plot = "2014")
}
