\name{simulate.fitContinuous}
\alias{simulate.fitContinuous}
\title{simulate method}
\usage{
  \method{simulate}{fitContinuous} (fit)
}
\value{
  simulated dataset
}
\description{
  simulate method
}
\details{
  rTraitCont might not be ideal here, should warn if alpha
  is large!  Currently only designed to simulate from the
  first fit
}

