% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geiger.R
\name{simulate.gfit}
\alias{simulate.gfit}
\title{simulate gfit}
\usage{
simulate.gfit(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{a gfit object}

\item{nsim}{number of sims}

\item{seed}{an optional seed for the simulations (not implemented)}

\item{...}{additional arguments, not implemented for gfit simulations}
}
\value{
simulated dataset
}
\description{
simulate method for gfit objects
}
