% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Slave_Lake}
\alias{Slave_Lake}
\title{Slave Lake Dataset}
\format{
A tibble (or data frame) with 730 rows and 5 variables:
\describe{
\item{Date}{Date of the observation (YYYY-MM-DD)}
\item{Tmin}{Minimum daily temperature (°C)}
\item{Tmax}{Maximum daily temperature (°C)}
\item{PPT}{Daily precipitation (cm)}
}
}
\source{
Data obtained from \href{https://climatedata.ca/download/#station-download}{ClimateData.ca}.
}
\usage{
data(Slave_Lake)
}
\description{
A dataset that includes daily weather observations collected from the Slave Lake RCS weather station, located at a latitude of 55.2931.
}
\keyword{datasets}
