\name{update.fitContinuous}
\alias{update.fitContinuous}
\title{update a fitContinuous model}
\usage{
  \method{update}{fitContinuous} (object, ...)
}
\arguments{
  \item{object}{the output of fitContinuous}

  \item{...}{data on which update will be based}
}
\value{
  the MLE estimated model for the given data
}
\description{
  update a fitContinuous model
}

