% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geiger.R
\name{update.gfit}
\alias{update.gfit}
\title{update gfit}
\usage{
update.gfit(object, ...)
}
\arguments{
\item{object}{a gfit object}

\item{...}{additional arguments, such as the data to use to update}
}
\value{
updated gfit object
}
\description{
update method for gfit objects
}

