% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{visuals}
\alias{visuals}
\title{Plots of selected parameters}
\usage{
visuals(
  input_data,
  var1,
  var2,
  var3,
  var4,
  x_label_vr1,
  x_label_vr2,
  x_label_vr3,
  x_label_vr4
)
}
\arguments{
\item{input_data}{A dataframe containing the variables to be plotted.}

\item{var1}{The variable to be plotted in the first panel.}

\item{var2}{The variable to be plotted in the second panel.}

\item{var3}{The variable to be plotted in the third panel.}

\item{var4}{The variable to be plotted in the fourth panel.}

\item{x_label_vr1}{The y-axis label for the first variable.}

\item{x_label_vr2}{The y-axis label for the second variable.}

\item{x_label_vr3}{The y-axis label for the third variable.}

\item{x_label_vr4}{The y-axis label for the fourth variable.}
}
\value{
An arrangement of multiple plots as a single, combined figure.
}
\description{
Plots of selected parameters
}
\examples{
input_data <- data.frame(Date = as.Date(c("2024-01-01", "2024-01-02",
"2023-01-03")), PET_Calculated = c(0.003444420, 0.015362293, 0.011119636),
Tmean = c(-31.55, -15.65, -19.00), PMC = c(10.005600, 9.686799, 9.688504),
eff_Precip_cm = c(0.00, 0.21, 0.01))
visuals(input_data,
var1="PET_Calculated",
var2="Tmean",
var3="PMC",
var4="eff_Precip_cm",
x_label_vr1 = "PET (cm)",
x_label_vr2 = "Tair (°C)",
x_label_vr3 = "PMC",
x_label_vr4 = "Eff PPT")
}
