% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_PET.R
\name{Calculate_PET}
\alias{Calculate_PET}
\title{Pryestly-Taylor Potential Evapotranspiration}
\usage{
Calculate_PET(
  input.data,
  latitude,
  alpha,
  y,
  Gsc,
  lambda,
  a,
  b,
  year_to_plot = NULL
)
}
\arguments{
\item{input.data}{A dataframe containing the columns 'Date', 'Tmin', 'Tmax', and 'PPT'.}

\item{latitude}{The location in decimal degrees.}

\item{alpha}{The Priestly-Taylor alpha constant.}

\item{y}{The psychrometric constant in kPa/°C.}

\item{Gsc}{The solar constant in MJ m^-2 min^-1.}

\item{lambda}{The latent heat of vaporization in MJ m^-3.}

\item{a}{A constant used in the calculation of shortwave radiation from Kext.}

\item{b}{A constant used in the calculation of shortwave radiation from Kext.}

\item{year_to_plot}{Select the year of choice to plot.}
}
\value{
A dataframe that retains the original columns and adds a column with the calculated Potential Evapotranspiration (PET).
}
\description{
Pryestly-Taylor Potential Evapotranspiration
}
\examples{
input_data <- data.frame(Date = as.Date(c("2024-01-01", "2024-01-02",
"2023-01-03")), Tmin = c(-36.1, -23.8, -25.6), Tmax = c(-27.0, -7.5, -12.4),
PPT = c(0.0, 2.1, 0.1))
Calculate_PET(input_data,
latitude = 56.9002499,
alpha = 1,
y = 0.063,
Gsc=0.0820,
lambda=2453,
a=0.11,
b=0.59)

}
