% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_PET.R
\name{Calculate_PET}
\alias{Calculate_PET}
\title{Pryestly-Taylor Potential Evapotranspiration}
\usage{
Calculate_PET(
  input_data,
  latitude,
  alpha,
  y,
  Gsc,
  lambda,
  a,
  b,
  year_to_plot = NULL
)
}
\arguments{
\item{input_data}{A data frame that must include 'Date', 'Tmin', and 'Tmax'. Tavg is optional.}

\item{latitude}{Latitude of the study site. \strong{User-dependent}; must be specified for correct solar radiation calculations.}

\item{alpha}{Priestly-Taylor coefficient. \strong{User-dependent}; typically ranges from 1 to 1.26 depending on local conditions.}

\item{y}{Psychrometric constant (kPa/°C). Default is 0.063. \strong{User-dependent} in some models.}

\item{Gsc}{Solar constant (MJ m\eqn{^{-2}} min\eqn{^{-1}}). Default is 0.0820.}

\item{lambda}{Latent heat of vaporization of water (MJ g\eqn{^{-3}}). Default is 2453. If not specified by the user, it will be internally calculated based on temperature.}

\item{a}{Constant in the Hargreaves calculation of Rn from extraterrestrial radiation (\eqn{K_\text{ext}}). Default is 0.17.}

\item{b}{Constant in the Hargreaves calculation of net radiation (Rn) from extraterrestrial radiation (\eqn{K_\text{ext}}). Default is 0.59.}

\item{year_to_plot}{Select the year of choice to plot.}
}
\value{
A dataframe that retains the original columns and adds a column with the calculated Potential Evapotranspiration (PET) in centimeters.
}
\description{
The Priestley-Taylor (PT) equation simplifies the Penman-Monteith method by combining the vapor deficit and convection terms into a single empirical constant, α.
}
\details{
PET is calculated as follows:

\deqn{ET_{pt} = \alpha \cdot \frac{\Delta \cdot R_n}{\lambda_v \cdot (\Delta + \gamma)} \cdot 100}{
ETpt = α * [Δ * Rn] / [λv * (Δ + γ)] * 100
}

Where:
\itemize{
\item \eqn{\alpha}{α} is the Priestley–Taylor coefficient (typically ~1),
\item \eqn{\Delta}{Δ} is the slope of the saturation vapor pressure curve (kPa/°C),
\item \eqn{R_n} is the net radiation (MJ/m²/day),
\item \eqn{\lambda_v}{λv} is the latent heat of vaporization (kJ/kg),
\item \eqn{\gamma} is the psychrometric constant (kPa/°C).
}

The result is returned in centimeters (cm) per day.
}
\examples{
input_data <- data.frame(Date = as.Date(c("2024-01-01", "2024-01-02",
"2023-01-03")), Tmin = c(-36.1, -23.8, -25.6), Tmax = c(-27.0, -7.5, -12.4),
PPT = c(0.0, 2.1, 0.1))
Calculate_PET(input_data,
latitude = 56.9002499,
alpha = 1,
y = 0.063,
Gsc=0.0820,
lambda=2453,
a=0.17,
b=0.59)

}
