% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{cleaning_data}
\alias{cleaning_data}
\title{Clean the initial dataframe}
\usage{
cleaning_data(input_data)
}
\arguments{
\item{input_data}{A dataframe that contains columns 'Date', 'Tmin', 'Tmax', 'PPT'}
}
\value{
A dataframe with date columns of class Date, and the other variables as numeric
}
\description{
Clean the initial dataframe
}
\examples{
input_data <- data.frame(Date = as.Date(c("2024-01-01", "2024-01-02",
"2023-01-03")), Tmin = c(-36.1, -23.8, -25.6), Tmax = c(-27.0, -7.5, -12.4),
PPT = c(0.0, 2.1, 0.1), ISI = c(1.1, 1.2, 1.5))
cleaning_data(input_data)
}
