\name{convert}
\alias{convert}

\title{
  Convert between ape and ouch trees
}
\description{
Given an ape tree (class phylo) returns an ouch tree 
(class ouchtree) or vice versa.  Returns regimes from the 
ouch fit object as well.  
}
\usage{
convert(ot)
}
\arguments{
  \item{ot}{
    a phylogenetic tree
}
}
\author{
  Carl Boettiger 
}


