% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_ppt.R
\name{effective_ppt}
\alias{effective_ppt}
\title{Effective precipitation}
\usage{
effective_ppt(input_data, column, threshold)
}
\arguments{
\item{input_data}{A dataframe that contains columns 'Date', 'Tmin', 'Tmax', 'PPT'}

\item{column}{Precipitation variable in centimeters}

\item{threshold}{A minimum amount of precipitation required before water reaches the peatland storage.}
}
\value{
A dataframe that includes both the original columns and the newly calculated effective precipitation values.
}
\description{
Effective precipitation
}
\examples{
input_data <- data.frame(Date = as.Date(c("2024-01-01", "2024-01-02",
"2023-01-03")), Tmin = c(-36.1, -23.8, -25.6), Tmax = c(-27.0, -7.5, -12.4),
PPT = c(0.0, 2.1, 0.1))
effective_ppt(input_data, column = "PPT", threshold = 0.01)
}
