% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fort_McMurray}
\alias{Fort_McMurray}
\title{Fort McMurray Dataset}
\format{
A tibble (or data frame) with 730 rows and 5 variables:
\describe{
\item{Date}{Date of the observation (YYYY-MM-DD)}
\item{Tmin}{Minimum daily temperature (°C)}
\item{Tmax}{Maximum daily temperature (°C)}
\item{PPT}{Daily precipitation (cm)}
}
}
\source{
Data obtained from \href{https://climatedata.ca/download/#station-download}{ClimateData.ca}.
}
\usage{
data(Fort_McMurray)
}
\description{
A dataset that includes daily weather observations collected from the Mildred Lake weather station in Alberta.
\strong{Location Coordinates:} Latitude: 56.6500, Longitude: -111.2167.
}
\keyword{datasets}
