% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMC_calc.R
\name{PMC_calc}
\alias{PMC_calc}
\title{PMC Calculation}
\usage{
PMC_calc(input_data, PET_column, A, B, start_PMC, C, Sy_min, PMC_min)
}
\arguments{
\item{input_data}{A dataframe containing the columns 'Date', 'PET_cm', and 'eff_Precip_cm'.}

\item{PET_column}{A PET column is necessary for calculating PCM. The required unit of measurement is centimeters.}

\item{A}{Equation paramater}

\item{B}{Equation parameter}

\item{start_PMC}{A parameter to establish the initial value for the water table depth in centimeters.}

\item{C}{Limited to a minimum of 0.3 to account for ET from trees}

\item{Sy_min}{Minimum Specific Yield (Sy)}

\item{PMC_min}{Minimum PMC value}
}
\value{
A dataframe that retains the original columns and adds a column with the calculated PMC.
}
\description{
PMC Calculation
}
\examples{
input_data <- data.frame(Date = as.Date(c("2024-01-01", "2024-01-02",
"2023-01-03")), eff_Precip_cm = c(0.0, 0.21, 0.01), PET_Calculated =
c(0.003444420, 0.015362293, 0.011119636))
PMC_calc(input_data, PET_column = "PET_Calculated",
A = 0.8558,
B = 0.0337,
start_PMC = 10,
C = 0.1,
Sy_min = 0.1,
PMC_min = -5)
}
