% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc.R
\name{pmc_fit}
\alias{pmc_fit}
\title{Fit any model used in PMC}
\usage{
pmc_fit(tree, data, model, ...)
}
\arguments{
\item{tree}{a phylogenetic tree. can be ouch or ape format}

\item{data}{trait data in ape or ouch format}

\item{model}{the name of the model to fit,}

\item{...}{whatever additional options would be provided 
to the model fit}
}
\value{
the object returned by the model fitting routine (gfit for geiger, hansen/brown for ouch)
}
\description{
The fitting function used by pmc to generalize fitting to both geiger and ouch models.
}

