% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{visuals}
\alias{visuals}
\title{Multi-panel Plot of Selected Variables}
\usage{
visuals(
  input_data,
  var1,
  var2,
  var3,
  var4,
  x_label_vr1,
  x_label_vr2,
  x_label_vr3,
  x_label_vr4,
  year_to_plot = NULL
)
}
\arguments{
\item{input_data}{A dataframe containing a 'Date' column and the variables to be plotted.}

\item{var1}{The name of the variable to be plotted in the first panel (as a string).}

\item{var2}{The name of the variable to be plotted in the second panel (as a string).}

\item{var3}{The name of the variable to be plotted in the third panel (as a string).}

\item{var4}{The name of the variable to be plotted in the fourth panel (as a string).}

\item{x_label_vr1}{The y-axis label for the first variable.}

\item{x_label_vr2}{The y-axis label for the second variable.}

\item{x_label_vr3}{The y-axis label for the third variable.}

\item{x_label_vr4}{The y-axis label for the fourth variable.}

\item{year_to_plot}{Optional. Numeric value indicating the year to display in a generated plot of daily PET.}
}
\value{
A combined figure containing four ggplot2 panels arranged vertically, each showing the time series of one variable.
}
\description{
This function creates a four-panel plot to visualize temporal trends in selected variables from a given dataset.
It is useful for comparing key hydrometeorological and modeled parameters over time.
}
\examples{
input_data <- data.frame(
  Date = as.Date(c("2024-01-01", "2024-01-02", "2024-01-03")),
  PET = c(0.0034, 0.0154, 0.0111),
  Tavg = c(-31.55, -15.65, -19.00),
  PMC = c(10.01, 9.69, 9.69),
  eff_Precip_cm = c(0.00, 0.21, 0.01)
)
visuals(
  input_data,
  var1 = "PET",
  var2 = "Tavg",
  var3 = "PMC",
  var4 = "eff_Precip_cm",
  x_label_vr1 = "PET (cm)",
  x_label_vr2 = "Tair (°C)",
  x_label_vr3 = "PMC",
  x_label_vr4 = "Effective Precipitation (cm)",
  year_to_plot = "2024"
)
}
